Add VmGroup class and refactor code to use it
[#157984461](https://www.pivotaltracker.com/story/show/157984461)
