+++ src/vsphere_cpi/lib/cloud/vsphere.rb
@@ -35,6 +35,7 @@ require 'cloud/vsphere/resources/task'
 require 'cloud/vsphere/drs_rules/drs_lock'
 require 'cloud/vsphere/drs_rules/drs_rule'
 require 'cloud/vsphere/drs_rules/vm_attribute_manager'
+require 'cloud/vsphere/drs_rules/vm_group'
 require 'cloud/vsphere/soap_stub'
 require 'cloud/vsphere/vm_creator'
 require 'cloud/vsphere/vm_provider'
