+++ src/vsphere_cpi/lib/cloud/vsphere/drs_rules/drs_lock.rb
@@ -10,9 +10,9 @@ module VSphereCloud
     class LockError < RuntimeError; end
     class TimeoutError < RuntimeError; end
 
-    def initialize(vm_attribute_manager, drs_lock_suffix='')
+    def initialize(vm_attribute_manager, drs_lock_name=DRS_LOCK_NAME)
       @vm_attribute_manager = vm_attribute_manager
-      @drs_lock_suffix = drs_lock_suffix
+      @drs_lock_name = drs_lock_name
     end
 
     def with_drs_lock
@@ -29,23 +29,19 @@ module VSphereCloud
 
     private
 
-    def drs_lock_name
-      DRS_LOCK_NAME + @drs_lock_suffix
-    end
-
     def acquire_lock
       # vm_attribute_manager.create will raise DuplicateName exception if that field already exists
       # Retrying until the call succeeds ensures that only one CPI process is modifying a DRS Group at a time
       retry_with_timeout do
-        @vm_attribute_manager.create(drs_lock_name)
+        @vm_attribute_manager.create(@drs_lock_name)
       end
     rescue TimeoutError
-      logger.debug("Failed to acquire drs lock: #{drs_lock_name}")
-      raise LockError.new("Failed to acquire DRS lock: #{drs_lock_name}")
+      logger.debug("Failed to acquire drs lock: #{@drs_lock_name}")
+      raise LockError.new("Failed to acquire DRS lock: #{@drs_lock_name}")
     end
 
     def release_lock
-      @vm_attribute_manager.delete(drs_lock_name)
+      @vm_attribute_manager.delete(@drs_lock_name)
     end
 
     def retry_with_timeout
