+++ src/vsphere_cpi/lib/cloud/vsphere/vm_creator.rb
@@ -208,17 +208,13 @@ module VSphereCloud
       drs_rule.add_vm(vm_mob)
     end
 
-    # using DrsRule here as this will eventually be used to create vm/host group rule
-    # for initial release we are only adding support for vm_group
     def add_vm_to_vm_group(vm_config, vm_mob, cluster)
       return if vm_config.vm_type.vm_group.nil?
-      drs_rule_name = DrsRule::DEFAULT_RULE_NAME #Passing it on purpose as this would be temporary
-      drs_rule = VSphereCloud::DrsRule.new(
-        drs_rule_name,
+      vm_group = VSphereCloud::VmGroup.new(
         @client,
         cluster.mob
       )
-      drs_rule.add_vm_to_vm_group(vm_mob, vm_config.vm_type.vm_group)
+      vm_group.add_vm_to_vm_group(vm_mob, vm_config.vm_type.vm_group)
     end
   end
 end
