+++ src/vsphere_cpi/lib/cloud/vsphere/vm_type.rb
@@ -13,7 +13,7 @@ module VSphereCloud
     end
 
     %w[cpu ram disk cpu_hot_add_enabled memory_hot_add_enabled nsx vmx_options
-       nsxt nested_hardware_virtualization upgrade_hw_version datacenters gpu vm_group ].each do |name|
+       nsxt nested_hardware_virtualization upgrade_hw_version datacenters vm_group ].each do |name|
       define_method(name) { @cloud_properties[name] }
     end
