+++ src/vsphere_cpi/spec/unit/cloud/vsphere/drs_rules/drs_rule_spec.rb
@@ -151,50 +151,4 @@ describe VSphereCloud::DrsRule, fake_logger: true do
     end
   end
 
-  describe '#add_vm_to_vm_group' do
-    let(:vm_1) { instance_double('VimSdk::Vim::VirtualMachine', name: 'vm-1') }
-    let(:vm_2) { instance_double('VimSdk::Vim::VirtualMachine', name: 'vm-2') }
-    let(:vm_group_name) {'vm-group'}
-    let(:vm_groups) { [ VimSdk::Vim::Cluster::VmGroup.new(name: vm_group_name, vm: [vm_1]) ] }
-    let(:configuration_ex) { double(:configuration_ex, group: vm_groups) }
-
-    before do
-      allow(datacenter_cluster).to receive(:configuration_ex).and_return(configuration_ex)
-      allow(datacenter_cluster).to receive(:reconfigure_ex).and_return(task)
-    end
-
-    context 'when vm_group exists' do
-      it 'adds vm to the existing vm_group with DRS lock' do
-        with_lock do
-          expect(datacenter_cluster).to receive(:reconfigure_ex) do |config|
-            group_spec = config.group_spec.first
-            expect(group_spec.operation).to eq(VimSdk::Vim::Option::ArrayUpdateSpec::Operation::EDIT)
-            group_info = group_spec.info
-            expect(group_info).to be_an_instance_of(VimSdk::Vim::Cluster::VmGroup)
-            expect(group_info.vm).to eq([vm_1, vm_2])
-            expect(group_info.name).to eq(vm_group_name)
-
-          end.ordered.and_return(task)
-        end
-        drs_rule.add_vm_to_vm_group(vm_2, vm_group_name)
-      end
-    end
-    context 'when vm_group does not exists' do
-      let(:vm_groups) { []}
-      it 'creates vm_group and adds vm to it with DRS lock' do
-        with_lock do
-          expect(datacenter_cluster).to receive(:reconfigure_ex) do |config|
-            group_spec = config.group_spec.first
-            expect(group_spec.operation).to eq(VimSdk::Vim::Option::ArrayUpdateSpec::Operation::ADD)
-            group_info = group_spec.info
-            expect(group_info).to be_an_instance_of(VimSdk::Vim::Cluster::VmGroup)
-            expect(group_info.vm).to eq([vm_1])
-            expect(group_info.name).to eq(vm_group_name)
-
-          end.ordered.and_return(task)
-        end
-        drs_rule.add_vm_to_vm_group(vm_1, vm_group_name)
-      end
-    end
-  end
 end
